requirements.txt
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
autopep8==1.6.0
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
colorama==0.4.5
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pydantic-sqlalchemy==0.0.9
PyJWT==2.4.0
PyMySQL==1.0.2
python-decouple==3.6
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
toml==0.10.2
typing_extensions==4.2.0
uvicorn==0.17.6
